fix: freeze when opening a live-app [LIVE-13411]
