+++ .changeset/tricky-forks-fly.md
@@ -1,0 +1,5 @@
+---
+"live-mobile": patch
+---
+
+fix: freeze when opening a live-app
