+++ apps/ledger-live-mobile/src/newArch/features/Web3Hub/screens/Web3HubApp/components/Web3Player/BottomBar.tsx
@@ -1,23 +1,18 @@
 import React, { RefObject, useCallback } from "react";
 import { TouchableOpacity } from "react-native";
 import { useTheme } from "@react-navigation/native";
-import { Trans } from "react-i18next";
 import Animated, {
   useAnimatedStyle,
   interpolate,
   Extrapolation,
   SharedValue,
 } from "react-native-reanimated";
-import { Flex, Text } from "@ledgerhq/native-ui";
+import { Flex } from "@ledgerhq/native-ui";
 import { AppManifest } from "@ledgerhq/live-common/wallet-api/types";
 import { ArrowLeftMedium, ArrowRightMedium, ReverseMedium } from "@ledgerhq/native-ui/assets/icons";
 import { safeGetRefValue, CurrentAccountHistDB } from "@ledgerhq/live-common/wallet-api/react";
-import { useDappCurrentAccount } from "@ledgerhq/live-common/wallet-api/useDappLogic";
 import { WebviewAPI, WebviewState } from "~/components/Web3AppWebview/types";
-import Button from "~/components/Button";
-import CircleCurrencyIcon from "~/components/CircleCurrencyIcon";
-import { useSelectAccount } from "~/components/Web3AppWebview/helpers";
-import { useMaybeAccountName } from "~/reducers/wallet";
+import SelectAccountButton from "./SelectAccountButton";
 
 const BAR_HEIGHT = 60;
 export const TOTAL_HEADER_HEIGHT = BAR_HEIGHT;
@@ -61,7 +56,6 @@ export function BottomBar({
   layoutY,
 }: BottomBarProps) {
   const { colors } = useTheme();
-  const { currentAccount } = useDappCurrentAccount(currentAccountHistDb);
   const shouldDisplaySelectAccount = !!manifest.dapp;
 
   const handleForward = useCallback(() => {
@@ -81,10 +75,6 @@ export function BottomBar({
 
     webview.reload();
   }, [webviewAPIRef]);
-
-  const { onSelectAccount } = useSelectAccount({ manifest, currentAccountHistDb });
-
-  const currentAccountName = useMaybeAccountName(currentAccount);
 
   const heightStyle = useAnimatedStyle(() => {
     if (!layoutY) return {};
@@ -132,27 +122,7 @@ export function BottomBar({
           </Flex>
 
           {shouldDisplaySelectAccount ? (
-            <Button type="primary" onPress={onSelectAccount}>
-              {!currentAccount ? (
-                <Text>
-                  <Trans i18nKey="common.selectAccount" />
-                </Text>
-              ) : (
-                <Flex flexDirection="row" height={50} alignItems="center" justifyContent="center">
-                  <CircleCurrencyIcon
-                    size={24}
-                    currency={
-                      currentAccount.type === "TokenAccount"
-                        ? currentAccount.token
-                        : currentAccount.currency
-                    }
-                  />
-                  <Text color={"neutral.c20"} ml={4}>
-                    {currentAccountName}
-                  </Text>
-                </Flex>
-              )}
-            </Button>
+            <SelectAccountButton manifest={manifest} currentAccountHistDb={currentAccountHistDb} />
           ) : null}
 
           <IconButton onPress={handleReload} alignSelf="flex-end">
